incorporated upstream's 2.1.4 release tarball minus the debian bits

+++ duplicity/util.py
@@ -19,71 +19,26 @@
 # along with duplicity; if not, write to the Free Software Foundation,
 # Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
 
-u"""
+"""
 Miscellaneous utilities.
 """
 
-from __future__ import print_function
-
-from future import standard_library
-standard_library.install_aliases()
-from builtins import isinstance
-from builtins import map
-from builtins import object
-from builtins import str
-
+import atexit
 import csv
 import errno
 import json
 import os
-import string
 import sys
 import traceback
-import atexit
+from io import StringIO
 
-if sys.version_info.major == 2:
-    from cStringIO import StringIO  # pylint: disable=import-error
-else:
-    from io import StringIO  # pylint: disable=import-error
-
-from duplicity import tarfile
 import duplicity.config as config
 import duplicity.log as log
-
-try:
-    # For paths, just use path.name/uname rather than converting with these
-    from os import fsencode, fsdecode  # pylint: disable=unused-import
-except ImportError:
-    # Most likely Python version < 3.2, so define our own fsencode/fsdecode.
-    # These are functions that encode/decode unicode paths to filesystem encoding,
-    # but the cleverness is that they handle non-unicode characters on Linux
-    # There is a *partial* backport to python available here:
-    # https://github.com/pjdelport/backports.os/blob/master/src/backports/os.py
-    # but if it cannot be trusted for full-circle translation, then we may as well
-    # just read and store the bytes version of the path as path.name before
-    # creating the unicode version (for path matching etc) and ensure that in
-    # real-world usage (as opposed to testing) we create the path objects from a
-    # bytes string.
-    # ToDo: Revisit this once we drop Python 2 support/the backport is complete
-
-    def fsencode(unicode_filename):
-        u"""Convert a unicode filename to a filename encoded in the system encoding"""
-        # For paths, just use path.name rather than converting with this
-        # If we are not doing any cleverness with non-unicode filename bytes,
-        # encoding to system encoding is good enough
-        return unicode_filename.encode(sys.getfilesystemencoding(), u"replace")
-
-    def fsdecode(bytes_filename):
-        u"""Convert a filename encoded in the system encoding to unicode"""
-        # For paths, just use path.uc_name rather than converting with this
-        # If we are not doing any cleverness with non-unicode filename bytes,
-        # decoding using system encoding is good enough. Use "ignore" as
-        # Linux paths can contain non-Unicode characters
-        return bytes_filename.decode(config.fsencoding, u"replace")
+from duplicity import tarfile
 
 
 def exception_traceback(limit=50):
-    u"""
+    """
     @return A string representation in typical Python format of the
             currently active/raised exception.
     """
@@ -92,33 +47,28 @@ def exception_traceback(limit=50):
     lines = traceback.format_tb(tb, limit)
     lines.extend(traceback.format_exception_only(type, value))
 
-    msg = u"Traceback (innermost last):\n"
-    if sys.version_info.major >= 3:
-        msg = msg + u"%-20s %s" % (str.join(u"", lines[:-1]), lines[-1])
-    else:
-        msg = msg + u"%-20s %s" % (string.join(lines[:-1], u""), lines[-1])
+    msg = "Traceback (innermost last):\n"
+    msg = msg + "%-20s %s" % (str.join("", lines[:-1]), lines[-1])
 
-    if sys.version_info.major < 3:
-        return msg.decode(u'unicode-escape', u'replace')
     return msg
 
 
 def escape(string):
-    u"""Convert a (bytes) filename to a format suitable for logging (quoted utf8)"""
-    string = fsdecode(string).encode(u'unicode-escape', u'replace')
-    return u"'%s'" % string.decode(u'utf8', u'replace').replace(u"'", u'\\x27')
+    """Convert a (bytes) filename to a format suitable for logging (quoted utf8)"""
+    string = os.fsdecode(string).encode("unicode-escape", "replace")
+    return "'%s'" % string.decode("utf8", "replace").replace("'", "\\x27")
 
 
 def uindex(index):
-    u"""Convert an index (a tuple of path parts) to unicode for printing"""
+    """Convert an index (a tuple of path parts) to unicode for printing"""
     if index:
-        return os.path.join(*list(map(fsdecode, index)))
+        return os.path.join(*list(map(os.fsdecode, index)))
     else:
-        return u'.'
+        return "."
 
 
 def uexc(e):
-    u"""Returns the exception message in Unicode"""
+    """Returns the exception message in Unicode"""
     # Exceptions in duplicity often have path names in them, which if they are
     # non-ascii will cause a UnicodeDecodeError when implicitly decoding to
     # unicode.  So we decode manually, using the filesystem encoding.
@@ -131,21 +81,16 @@ def uexc(e):
                 return m
             elif isinstance(m, bytes):
                 # Encoded, likely in filesystem encoding
-                return fsdecode(m)
+                return os.fsdecode(m)
         # If the function did not return yet, we did not
         # succeed in finding a string; return the whole message.
-        # This fails for Python 2, so only do this in Python 3.
-        if sys.version_info[0] > 2:
-            return str(e)
-        # For Python 2, fall back to returning an empty string.
-        else:
-            return u''
+        return str(e)
     else:
-        return u''
+        return ""
 
 
 def maybe_ignore_errors(fn):
-    u"""
+    """
     Execute fn. If the global configuration setting ignore_errors is
     set to True, catch errors and log them but do continue (and return
     None).
@@ -157,15 +102,13 @@ def maybe_ignore_errors(fn):
         return fn()
     except Exception as e:
         if config.ignore_errors:
-            log.Warn(_(u"IGNORED_ERROR: Warning: ignoring error as requested: %s: %s")
-                     % (e.__class__.__name__, uexc(e)))
+            log.Warn(_("IGNORED_ERROR: Warning: ignoring error as requested: %s: %s") % (e.__class__.__name__, uexc(e)))
             return None
         else:
             raise
 
 
 class BlackHoleList(list):
-
     def append(self, x):
         pass
 
@@ -185,7 +128,7 @@ def make_tarfile(mode, fp):
     # yet.  So we want to ignore ReadError exceptions, which are used to signal
     # this.
     try:
-        tf = tarfile.TarFile(u"arbitrary", mode, fp)
+        tf = tarfile.TarFile("arbitrary", mode, fp)
         # Now we cause TarFile to not cache TarInfo objects.  It would end up
         # consuming a lot of memory over the lifetime of our long-lasting
         # signature files otherwise.
@@ -200,13 +143,13 @@ def get_tarinfo_name(ti):
     # and later ensure they they *don't* have /.  ::shrug::  Internally, we
     # continue to use pre-2.6 method.
     if ti.isdir() and not ti.name.endswith(r"/"):
-        return ti.name + r"/"
+        return f"{ti.name}/"
     else:
         return ti.name
 
 
 def ignore_missing(fn, filename):
-    u"""
+    """
     Execute fn on filename.  Ignore ENOENT errors, otherwise raise exception.
 
     @param fn: callable
@@ -223,20 +166,20 @@ def ignore_missing(fn, filename):
 
 @atexit.register
 def release_lockfile():
-    if config.lockfile:
-        log.Debug(_(u"Releasing lockfile %s") % config.lockpath)
+    if config.lockfile and os.path.exists(config.lockpath):
+        log.Debug(_("Releasing lockfile %s") % os.fsdecode(config.lockpath))
         try:
             config.lockfile.release()
-            config.lockfile = None
             os.remove(config.lockpath)
-            config.lockpath = u""
+            config.lockfile = None
+            config.lockpath = ""
         except Exception as e:
-            log.Error(u"Could not release lockfile: %s" % str(e))
+            log.Error(f"Could not release lockfile: {str(e)}")
             pass
 
 
 def copyfileobj(infp, outfp, byte_count=-1):
-    u"""Copy byte_count bytes from infp to outfp, or all if byte_count < 0
+    """Copy byte_count bytes from infp to outfp, or all if byte_count < 0
 
     Returns the number of bytes actually written (may be less than
     byte_count if find eof.  Does not close either fileobj.
@@ -245,7 +188,7 @@ def copyfileobj(infp, outfp, byte_count=-1):
     blocksize = 64 * 1024
     bytes_written = 0
     if byte_count < 0:
-        while 1:
+        while True:
             buf = infp.read(blocksize)
             if not buf:
                 break
@@ -265,7 +208,7 @@ def copyfileobj(infp, outfp, byte_count=-1):
 
 
 def which(program):
-    u"""
+    """
     Return absolute path for program name.
     Returns None if program not found.
     """
@@ -278,8 +221,8 @@ def which(program):
         if is_exe(program):
             return program
     else:
-        for path in os.getenv(u"PATH").split(os.pathsep):
-            path = path.strip(u'"')
+        for path in os.getenv("PATH").split(os.pathsep):
+            path = path.strip('"')
             exe_file = os.path.abspath(os.path.join(path, program))
             if is_exe(exe_file):
                 return exe_file
@@ -288,34 +231,59 @@ def which(program):
 
 
 def start_debugger():
-    if not os.getenv(u'DEBUG_RUNNING', None) and (u'--pydevd' in sys.argv or os.getenv(u'PYDEVD', None)):
+    if "--pydevd" in sys.argv or os.environ.get("PYDEVD", None):
         try:
-            import pydevd_pycharm as pydevd  # pylint: disable=import-error
+            import pydevd_pycharm  # pylint: disable=import-error
         except ImportError:
-            try:
-                import pydevd  # pylint: disable=import-error
-            except ImportError:
-                log.FatalError(u"Module pydevd_pycharm or pydevd must be available for debugging.\n"
-                               u"Remove '--pydevd' from command line and PYDEVD from environment\n"
-                               u"to avoid activating the debugger.")
+            log.FatalError(
+                "Module pydevd_pycharm must be available for debugging.\n"
+                "Remove '--pydevd' from command line and unset 'PYDEVD'\n"
+                "from the environment to avoid starting the debugger."
+            )
 
+        # NOTE: this needs to be customized for your system
+        debug_host = "dione.local"
+        debug_port = 6700
+
+        # get previous pid:port if any
+        # return if pid the same as ours
+        prev_port = None
+        debug_running = os.environ.get("DEBUG_RUNNING", False)
+        if debug_running:
+            prev_pid, prev_port = list(map(int, debug_running.split(":")))
+            if prev_pid == os.getpid():
+                return
+
+        # new pid, next port, start a new debugger
+        if prev_port:
+            debug_port = int(prev_port) + 1
+
+        # ignition
         try:
-            # NOTE: this needs to be customized for your system
-            pydevd.settrace(u'dione.local', port=6789, stdoutToServer=True, stderrToServer=True)
+            pydevd_pycharm.settrace(
+                debug_host,
+                port=debug_port,
+                suspend=False,
+                stdoutToServer=True,
+                stderrToServer=True,
+                # patch_multiprocessing=True,
+            )
+            log.Info(f"Connection {debug_host}:{debug_port} accepted for debug.")
         except ConnectionRefusedError as e:
-            log.FatalError(u"Connection refused for debug.  Check your setup.")
+            log.Info(f"Connection {debug_host}:{debug_port} refused for debug: {str(e)}")
 
-        # In a dev environment the path is screwed so fix it.
+        # in a dev environment the path is screwed so fix it.
         base = sys.path.pop(0)
         base = base.split(os.path.sep)[:-1]
         base = os.path.sep.join(base)
         sys.path.insert(0, base)
 
-        os.environ[u'DEBUG_RUNNING'] = u'yes'
+        # save last debug pid:port used
+        os.environ["DEBUG_RUNNING"] = f"{os.getpid()}:{debug_port}"
 
 
 def merge_dicts(*dict_args):
-    u"""
+    """
     Given any number of dictionaries, shallow copy and merge into a new dict,
     precedence goes to key-value pairs in latter dictionaries.
     """
@@ -326,7 +294,7 @@ def merge_dicts(*dict_args):
 
 
 def csv_args_to_dict(arg):
-    u"""
+    """
     Given the string arg in single line csv format, split into pairs (key, val)
     and produce a dictionary from those key:val pairs.
     """
@@ -339,14 +307,12 @@ def csv_args_to_dict(arg):
     return mydict
 
 
-# TODO: just use util.fsdecode().casefold() directly when python27 is gone
-def casefold_compat(s):
-    u"""
-    Compatability function for casefolding which provides an acceptable for
-    older pythons. Can likely be removed once python2 support is no longer o
-    any interest.
+class BytesEncoder(json.JSONEncoder):
     """
-    if sys.version_info.major >= 3 and sys.version_info.minor >= 3:
-        return s.casefold()
-    else:
-        return s.lower()
+    JSON doesn't allow byte type values. Converting them to unicode strings
+    """
+
+    def default(self, obj):
+        if isinstance(obj, bytes):
+            return obj.decode()
+        return json.JSONEncoder.default(self, obj)
